requirements.txt
aiohttp>=3.9.0
Pillow>=10.0.0
pydantic>=2.0.0
